+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 <!-- markdownlint-disable first-line-h1 -->
+## 3.12.3
+
+*2026-03-27*
+
+**Added**
+
+- Content inside `<math>`, `<chem>` and `<ce>` tags in the MediaWiki mode is now syntax highlighted
+
 ## 3.12.2
 
 *2026-03-26*